Added all sensor reading functions and fixed Ignition code checks
Conversions still needed on all sensors but Anemometer

+++ Application/Inc/misc/datastructs.h
@@ -24,7 +24,7 @@ typedef struct
 	uint8_t main_ignition_state;
 	uint8_t sec_ignition_state;
 	uint8_t hose_disconnect_state;
-	uint32_t code;	//NOT BEING SENT TO GST
+	uint8_t code;	//NOT BEING SENT TO GST
 	float battery_level;
 	float hose_pressure;
 	float hose_temperature;
